Corrected significant error in the ASet.Delete method
In fact, some cases were wrongly implemented breaking the ASet invariant
and leading to inconsistent sets on element deletion.
This have let the SOAG generator fail for the single-sweep.eag grammar example.
Additionally, added extensive unit tests for ASet.

This correction has been redone on branch 0.2.1 as agreed with @linkrope after having already been done on master as preparation to get rid of ASets at the master (replacing by a equally efficient D-Set implementation) and therefore to preserve the implementation idea (with correction) in this branch.