Corrected logging of anonymous nonterminals.
In  fact, SOAG.Sym[] may contain anonymous nonterninal which are not in use or part of computation. Their .FirstOcc is -1, so the array access fails. However, the EAG.HNontRepr() can deal with that. So we are using it here instead. And, it gets easier.

Taken over from https://github.com/linkrope/epsilon/pull/3 as we repeat the changes at this branch.